% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_se.R
\name{add_se}
\alias{add_se}
\title{add_se}
\usage{
add_se(model, name_f, name_x = "Intercept")
}
\arguments{
\item{model:}{a lm or glm fitted model}

\item{name_f:}{a character string with the name of the categorical (factor) variable}

\item{name_x:}{a character string with the name fo the interacting variable, by default the intercept}
}
\description{
From 2017 analysis. Preparation for xtable.
}
\details{
I nicked this somewhere, don't remember where
}
