% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInd.R
\name{calcInd}
\alias{calcInd}
\title{calcInd}
\usage{
calcInd(
  obs = NULL,
  exp = NULL,
  type = NULL,
  region_short = NULL,
  habitat = NULL,
  year = NULL,
  aWeight = amountWeights,
  dWeight = diffWeights,
  nIter = 9,
  save.draws = T,
  digits = 7
)
}
\arguments{
\item{obs}{Optional. Dataframe or tibble of observation community. Typically gotten from}

\item{exp}{Optional. Expectation community. Typically gotten from}

\item{type}{Needed if obs and exp not provided. Type of taxa (bumblebees, butterflies)}

\item{region_short}{Needed if obs and exp not provided. Shortname of region (sor, trond, ost).}

\item{habitat}{Needed if obs and exp not provided. Type of habitat (skogsmark, gressmark).}

\item{year}{Needed if obs not provided. Year of observations.}

\item{aWeight}{Named vector of weights for amounts. Relative weights for common, average and rare species. Typically gotten from}

\item{nIter}{Numeric. How many replicate draws for the uncertainty calculations. Set 0 or NULL for no uncertainty estimation.}

\item{save.draws}{Boolean. Should the draws be saved}

\item{digits}{Integer. How many digits of precision to use.}

\item{dWeights}{Named matrix of differnce weights. Relative weights for differences between expectation and observation communities.
Typically gotten from}
}
\description{
Calculate indicator values, with uncertainty
}
