% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComplData.R
\name{getComplData}
\alias{getComplData}
\title{getComplData}
\usage{
getComplData(
  type = c("Bumblebees", "Butterflies"),
  region_short = c("Trond", "Ost", "Sor", "Vest", "Nord"),
  habitat = c("Gressmark", "Skogsmark"),
  year = 2010,
  dataConnection = "con"
)
}
\arguments{
\item{type}{What data to get, "Bumblebees" or "Butterflies"}

\item{region_short}{Get data from what region? "Trond", "Ost", "Sor", "Vest", "Nord"}

\item{habitat}{Get data from what habitat type? Gressmark or Skogsmark}

\item{year}{Get data from what year?}
}
\description{
Get observation data from complete surveys of butterflies and bumblebees from the database
}
\details{
This is suitable for calculating the Nature Index indicators
}
