% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTransectWalks.R
\name{getTransectWalks}
\alias{getTransectWalks}
\title{getTransectWalks}
\usage{
getTransectWalks(include_id = FALSE)
}
\arguments{
\item{include_id}{}
}
\value{
A tibble of transect walks
}
\description{
Retrieve date and time for all recorded transect walks.
}
\examples{

\dontrun{

humleSommerfuglConnect
humlesommerfConnect()

all_transect_walks <- getTransectWalks()
}

}
