% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indPlot.R
\name{indPlot}
\alias{indPlot}
\title{indPlot}
\usage{
indPlot(input = NULL, confLevel = "90\%", errorWidth = 0.2,
  lineWidth = 2, ylab = "Indeks", xlab = "År", dodge = 0.1,
  pointSize = 0.2, errorColour = "black", errorbarAlpha = 0.4)
}
\arguments{
\item{input}{data object, from "collectIndData"}

\item{confLevel}{Confidence level of error bars. "90\%" or "95\%".}

\item{errorWidth}{widht of error bars}

\item{lineWidth}{width of lines}

\item{ylab}{default = "Indeks"}

\item{xlab}{default "År"}

\item{dodge}{How much should the points dodge? Default =  0.1}

\item{pointSize}{default = 0.2}

\item{errorColour}{color of errorbars. Default =  "black"}
}
\description{
Plot bumblebee and butterfly indicators
}
\details{
This function assumes an input object from the function "collectIndData"
}
