% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comm_index.R
\name{plot.comm_index}
\alias{plot.comm_index}
\title{plot.comm_index Plot function for comm_index class}
\usage{
\method{plot}{comm_index}(x, ...)
}
\arguments{
\item{x}{a comm_index object}

\item{...}{}
}
\value{
a ggplot object
}
\description{
Just an ugly stub at this point
}
\examples{
\dontrun{
plot(sInd2022OstGress)
}
}
