% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotArt.R
\name{plotArt}
\alias{plotArt}
\title{plotArt}
\usage{
plotArt(obs, exp, new.names = NULL, ...)
}
\arguments{
\item{obs}{observed community}

\item{exp}{expected community}

\item{new.names}{placeholder (fix)}
}
\description{
Plot bumblebee and butterfly occurrences in the shiny app style
}
