% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_comm_index.R
\name{print.comm_index}
\alias{print.comm_index}
\title{print.comm_index Print function for comm_index class
Bare-bones as of now}
\usage{
\method{print}{comm_index}(x, ...)
}
\arguments{
\item{x}{an object of comm_index class}

\item{...}{}
}
\value{
a named list
}
\description{
print.comm_index Print function for comm_index class
Bare-bones as of now
}
\examples{

\dontrun{

print(sInd2022OstGress)

}
}
