% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshViews.R
\name{refreshViews}
\alias{refreshViews}
\title{refreshViews}
\usage{
refreshViews()
}
\description{
Refresh the materialized views
}
\details{
Use this after each insert or update of the basic data. You need to be owner of parent tables.
}
