% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sInd2022TrondGress}
\alias{sInd2022TrondGress}
\title{Bumblebee and butterfly indicator objects}
\format{
An object of class \code{comm_index} (inherits from \code{list}) of length 8.
}
\usage{
sInd2022TrondGress
}
\description{
All indicies from 2009 to 2018 based on 9999 bootstraps
}
\keyword{datasets}
