% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyBarplot.R
\name{surveyBarplot}
\alias{surveyBarplot}
\title{surveyBarplot}
\usage{
surveyBarplot(input, ylab = "", xlab = "Antall")
}
\arguments{
\item{input}{An input from getAllData}

\item{ylab}{character, default = "Art".}

\item{xlab}{character, default = "Antall"}
}
\description{
Barplot of total number of survey records, divided by region
}
