% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectIndData.R
\name{collectIndData}
\alias{collectIndData}
\title{collectIndData}
\usage{
collectIndData(taxa = c("bumblebees", "butterflies"),
  habitat = c("gressmark", "skogsmark"), years = 2009:2018)
}
\arguments{
\item{taxa}{bumblees or butterflies}

\item{years}{what years to collect?}

\item{habita}{gressmark or skogsmark}
}
\value{
a tibble with indicator point estimates and lower and upper uncertainty bounds.
}
\description{
Bundle indicator values into a form that is easy to plot
}
\details{
This function assumes that the environment holds a bunch of bumblebee or butterfly indicator objects with a specific naming format (e.g. hInd2014SorGress)
}
