% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectIndData.R
\name{collectIndData}
\alias{collectIndData}
\title{collectIndData}
\usage{
collectIndData(
  type = c("Bumblebees", "Butterflies"),
  habitat = c("Gressmark", "Skogsmark"),
  years = 2009:2018
)
}
\arguments{
\item{years}{what years to collect?}

\item{taxa}{bumblees or butterflies}

\item{habita}{gressmark or skogsmark}
}
\value{
a tibble with indicator point estimates and lower and upper uncertainty bounds.
}
\description{
Bundle indicator values into a form that is easy to plot
}
\details{
This function assumes that the environment holds a bunch of bumblebee or butterfly indicator objects with a specific naming format (e.g. hInd2014SorGress)
}
\examples{

\dontrun{

data("hInd2010OstGress")
data("hInd2011OstGress")

tt <- collectIndData(type = "Bumblebees",
                     habitat = "Gressmark",
                     years = 2010:2011)

}


}
