% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllData.R
\name{getAllData}
\alias{getAllData}
\title{getAllData}
\usage{
getAllData(
  type = NULL,
  habitat = NULL,
  year = 2009:2018,
  dataConnection = "con",
  language = "norsk"
)
}
\arguments{
\item{type}{bumblebees or butterflies}

\item{habitat}{gressmark or skogsmark}

\item{year}{which year to get data from. Single year or span of years}

\item{dataConnection}{name of the database connection}
}
\description{
Get every record of bumblebees or butterflies from the database
}
