% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllData.R
\name{getAllData}
\alias{getAllData}
\title{getAllData}
\usage{
getAllData(
  type = c("Bumblebees", "Butterflies"),
  habitat = c("Gressmark", "Skogsmark"),
  year = 2009:2018,
  dataConnection = "con",
  language = "Norsk"
)
}
\arguments{
\item{type}{What data to get, "Bumblebees" or "Butterflies"}

\item{habitat}{Get data from what habitat type? Gressmark or Skogsmark}

\item{year}{which year to get data from. Single year or span of years}

\item{dataConnection}{name of the database connection}

\item{language}{What language for the species names? "Latin" or "Norsk"}
}
\description{
Get every record of bumblebees or butterflies from the database
}
