% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpValues.R
\name{getExpValues}
\alias{getExpValues}
\title{getExpValues}
\usage{
getExpValues(
  type = c("Bumblebees", "Butterflies"),
  region_short = c("Trond", "Ost", "Sor", "Vest"),
  habitat = c("Gressmark", "Skogsmark")
)
}
\description{
Get expectation communities for the NI indicator calculations from the database.
}
