% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpValues.R
\name{getExpValues}
\alias{getExpValues}
\title{getExpValues}
\usage{
getExpValues(
  type = c("bumblebees", "butterflies"),
  region_short = c("trond", "ost", "sor"),
  habitat = c("gressmark", "skogsmark")
)
}
\description{
Get expectation communities for the NI indicator calculations from the database.
}
