% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humlesommerfConnect.R
\name{humlesommerfConnect}
\alias{humlesommerfConnect}
\title{humlesommerfConnect}
\usage{
humlesommerfConnect(
  username = NULL,
  password = NULL,
  host = "T2lippgsql02.nina.no",
  dbname = "humlesommerf_refact",
  connectionName = "con",
  ...
)
}
\description{
Connect to the humlesommerf database at ninradardata01
}
