% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humlesommerfConnect.R
\name{humlesommerfConnect}
\alias{humlesommerfConnect}
\title{humlesommerfConnect}
\usage{
humlesommerfConnect(username = "humlereader", password = "naturindeks",
  host = "ninpgsql02.nina.no", dbname = "humlesommerf",
  connectionName = "con", bigint = "integer", ...)
}
\description{
Connect to the humlesommerf database at ninpgsql02
}
