% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotArt.R
\name{plotArt}
\alias{plotArt}
\title{plotArt}
\usage{
plotArt(obs, exp, addZeroes = FALSE, addLegend = FALSE, ...)
}
\arguments{
\item{obs}{observed community (Its species names will be used)}

\item{exp}{expected community}

\item{addZeroes}{Add non-observed species, boolean}

\item{addLegend}{Add legend, boolean}
}
\description{
Plot bumblebee and butterfly occurrences in the shiny app style
}
