% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotArt.R
\name{plotArt}
\alias{plotArt}
\title{plotArt}
\usage{
plotArt(obs, exp, ...)
}
\arguments{
\item{obs}{observed community (Its species names will be used)}

\item{exp}{expected community}
}
\description{
Plot bumblebee and butterfly occurrences in the shiny app style
}
